requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-multipart
google-generativeai
python-dotenv
faiss-cpu
imblearn
requests
streamlit
